+++ lib/core/modules/coderunner/codeRunner.js
@@ -1,7 +1,4 @@
 const RunCode = require('./runCode.js');
-const EmbarkJS = require('embarkjs');
-const IpfsApi = require('ipfs-api');
-const Web3 = require('web3');
 
 class CodeRunner {
   constructor(options) {
@@ -16,7 +13,7 @@ class CodeRunner {
     this.IpcClientListen();
     this.registerEvents();
     this.registerCommands();
-    this.registerEmbarkJs();
+    this.events.emit('runcode:ready');
   }
 
   registerIpcEvents() {
@@ -55,23 +52,6 @@ class CodeRunner {
     this.events.setCommandHandler('runcode:eval', this.evalCode.bind(this));
   }
 
-  registerEmbarkJs() {
-    this.registerVar('IpfsApi', IpfsApi, false);
-    this.registerVar('Web3', Web3, false);
-    this.registerVar('EmbarkJS', EmbarkJS, false);
-
-    if (this.ipc.connected) {
-      return;
-    }
-    this.events.on('code-generator-ready', () => {
-      this.events.request('code-generator:embarkjs:provider-code', (code) => {
-        const func = () => {};
-        this.evalCode(code, func, true);
-        this.evalCode(this.getInitProviderCode(), func, true);
-      });
-    });
-  }
-
   registerVar(varName, code, toRecord = true) {
     if (this.ipc.isServer() && toRecord) {
       this.commands.push({varName, code});
@@ -81,9 +61,7 @@ class CodeRunner {
   }
 
   evalCode(code, cb, forConsoleOnly = false) {
-    if (!cb) {
-      cb = function() {};
-    }
+    cb = cb || function() {};
     const awaitIdx = code.indexOf('await');
     if (awaitIdx > -1) {
       if (awaitIdx < 2) {
@@ -108,27 +86,6 @@ class CodeRunner {
     }
 
     cb(null, result);
-  }
-
-  getInitProviderCode() {
-    const codeTypes = {
-      'communication': this.config.communicationConfig || {},
-      'names': this.config.namesystemConfig || {},
-      'storage': this.config.storageConfig || {}
-    };
-
-    return this.plugins.getPluginsFor('initConsoleCode').reduce((acc, plugin) => {
-      Object.keys(codeTypes).forEach(codeTypeName => {
-        (plugin.embarkjs_init_console_code[codeTypeName] || []).forEach(initCode => {
-          let [block, shouldInit] = initCode;
-          if (shouldInit.call(plugin, codeTypes[codeTypeName])) {
-            acc += block;
-          }
-        });
-      });
-      return acc;
-    }, '');
-
   }
 }
